fix(wix-vibe-headless): split oversized reference files to stay under 10k chars
All 8 JS helper files exceeded the ~10k char fetch limit that some agents apply,
causing functions (including cart, checkout, booking) to be silently truncated.

Splits:
- wix-restaurants.js (28k) → wix-restaurants-menu.js + wix-restaurants-ordering.js + wix-restaurants-reservations.js
- wix-events.js (22k) → wix-events-browse.js + wix-events-registration.js
- wix-store.js (18k) → wix-store-catalog.js + wix-store-cart.js
- wix-bookings.js (15k) → wix-bookings-services.js + wix-bookings-checkout.js

Trims verbose inline data-model comment blocks in blog, cms, pricing-plans,
and portfolio (single-file verticals) to bring them under 10k.

Updates INSTRUCTIONS.md for all 4 split verticals and the SKILL.md routing
table to reflect the new file structure. Adds scripts/check-truncation.mjs
as an ongoing diagnostic tool.
