+++ skills/wix-vibe-headless/references/bookings/wix-bookings-checkout.js
@@ -1,0 +1,128 @@
+import { wixApiRequest } from "./wix-client.js";
+
+// Data model reference: see INSTRUCTIONS.md
+// Service and TimeSlot shapes: see wix-bookings-services.js
+
+// Wix Bookings app id — required inside catalogReference for the eCommerce checkout.
+const BOOKINGS_APP_ID = "13d21c63-b5ec-5912-8397-c3a5ddb27a97";
+
+/**
+ * Wix Bookings Booking (Bookings Writer V2) — what createBooking returns.
+ * Full model: https://dev.wix.com/docs/api-reference/business-solutions/bookings/bookings/bookings-writer-v2/create-booking.md
+ *
+ *   id {string}, status "CREATED" (not yet on calendar; confirmed after hosted checkout),
+ *   paymentStatus {string} — "NOT_PAID" until checkout completes,
+ *   bookedEntity.slot {object} — { serviceId, scheduleId, startDate, endDate, timezone, location },
+ *   contactDetails {object}, totalParticipants {number}
+ */
+
+/** Visitor's local IANA time zone — used as default for the booking timezone. */
+function defaultTimeZone() {
+  try {
+    return Intl.DateTimeFormat().resolvedOptions().timeZone || "UTC";
+  } catch {
+    return "UTC";
+  }
+}
+
+/**
+ * Create a booking for an appointment slot. The booking starts as status "CREATED" and is NOT
+ * yet on the calendar — it is confirmed automatically after the buyer completes the hosted
+ * checkout (call checkoutBooking next). selectedPaymentOption is "ONLINE".
+ *
+ * Pass a slot from listAvailableSlots()/getAvailableSlot() — it carries serviceId, scheduleId,
+ * localStartDate, localEndDate, location, availableResources. Always call getAvailableSlot
+ * first to re-validate and get staff info. Throws on an unbookable slot or missing booking id.
+ * https://dev.wix.com/docs/api-reference/business-solutions/bookings/bookings/bookings-writer-v2/create-booking.md
+ *
+ * @param {object} slot                       A TimeSlot from listAvailableSlots/getAvailableSlot.
+ * @param {{ firstName?: string, lastName?: string, email: string, phone?: string }} contactDetails
+ * @param {{ totalParticipants?: number, timeZone?: string, title?: string }} [options]
+ * @returns {Promise<object>} The created booking (status "CREATED").
+ */
+export async function createBooking(slot, contactDetails, { totalParticipants = 1, timeZone, title } = {}) {
+  if (!slot || slot.bookable === false) {
+    throw new Error("Cannot book: the selected slot is not bookable. Re-check availability and pick another time.");
+  }
+  if (!slot.serviceId || !slot.scheduleId || !slot.localStartDate || !slot.localEndDate) {
+    throw new Error("Cannot book: slot is missing serviceId/scheduleId/localStartDate/localEndDate.");
+  }
+  if (!contactDetails?.email) {
+    throw new Error("Cannot book: contactDetails.email is required.");
+  }
+
+  const resource = slot.availableResources?.[0]?.resources?.[0];
+
+  const bookedSlot = {
+    serviceId: slot.serviceId,
+    scheduleId: slot.scheduleId,
+    startDate: slot.localStartDate,
+    endDate: slot.localEndDate,
+    timezone: timeZone || defaultTimeZone(),
+    ...(slot.location ? { location: slot.location } : {}),
+    ...(resource ? { resource: { id: resource.id, name: resource.name } } : {}),
+  };
+
+  const res = await wixApiRequest("/bookings/v2/bookings", {
+    method: "POST",
+    body: {
+      booking: {
+        bookedEntity: { slot: bookedSlot, title: title || undefined, tags: ["INDIVIDUAL"] },
+        contactDetails,
+        additionalFields: [],
+        totalParticipants,
+        selectedPaymentOption: "ONLINE",
+      },
+      participantNotification: { notifyParticipants: true },
+    },
+  });
+  const booking = res?.booking;
+  if (!booking?.id) throw new Error("Failed to create the booking (no booking id returned).");
+  return booking;
+}
+
+/**
+ * Create an eCommerce checkout for a created booking and return the hosted checkout URL.
+ * Redirect the buyer there (window.location.href = ...). On return, the booking is confirmed.
+ * Throws if no redirect URL is produced.
+ * https://dev.wix.com/docs/rest/business-solutions/e-commerce/checkout/create-checkout.md
+ * @param {string} bookingId  booking.id from createBooking().
+ * @returns {Promise<string>} The hosted-checkout URL to redirect to.
+ */
+export async function checkoutBooking(bookingId) {
+  if (!bookingId) throw new Error("checkoutBooking requires a bookingId.");
+
+  const checkoutRes = await wixApiRequest("/ecom/v1/checkouts", {
+    method: "POST",
+    body: {
+      channelType: "WEB",
+      lineItems: [
+        { quantity: 1, catalogReference: { appId: BOOKINGS_APP_ID, catalogItemId: bookingId } },
+      ],
+    },
+  });
+  const checkoutId = checkoutRes?.checkout?.id;
+  if (!checkoutId) throw new Error("Failed to create a checkout for the booking.");
+
+  const redirect = await wixApiRequest("/headless/v1/redirect-session", {
+    method: "POST",
+    body: { ecomCheckout: { checkoutId }, callbacks: { postFlowUrl: window.location.href } },
+  });
+  const url = redirect?.redirectSession?.fullUrl;
+  if (!url) throw new Error("Failed to create the checkout redirect session.");
+  return url;
+}
+
+/**
+ * Convenience: create the booking and return the hosted-checkout URL in one call.
+ * Equivalent to createBooking() then checkoutBooking(booking.id). Throws loudly on any failure.
+ * @param {object} slot
+ * @param {{ firstName?: string, lastName?: string, email: string, phone?: string }} contactDetails
+ * @param {{ totalParticipants?: number, timeZone?: string, title?: string }} [options]
+ * @returns {Promise<{ booking: object, checkoutUrl: string }>}
+ */
+export async function bookAndCheckout(slot, contactDetails, options = {}) {
+  const booking = await createBooking(slot, contactDetails, options);
+  const checkoutUrl = await checkoutBooking(booking.id);
+  return { booking, checkoutUrl };
+}
