+++ skills/wix-vibe-headless/references/restaurants/INSTRUCTIONS.md
@@ -1,7 +1,13 @@
 
 # Wix Restaurants Skill
 
-> **Source files (in this skill):** the shared transport `references/shared/wix-client.js` and this vertical's `references/restaurants/wix-restaurants.js`. Copy **both** into your app's `src/rest/` side by side — the helper does `import { wixApiRequest } from "./wix-client.js"`, so they must land in the same folder.
+> **Source files (in this skill):** the shared transport `references/shared/wix-client.js` and the helper file(s) you need from `references/restaurants/`. All helpers import from `"./wix-client.js"`, so copy them into the same folder (e.g. `src/rest/`).
+>
+> | Need | Copy |
+> |---|---|
+> | Menu display (always) | `wix-restaurants-menu.js` |
+> | Online ordering (cart + checkout) | `wix-restaurants-ordering.js` |
+> | Table reservations | `wix-restaurants-reservations.js` |
 
 Builds a real, client-only Wix restaurant experience. The browser talks to Wix directly over a
 public `WIX_CLIENT_ID`. Never mock the menu; never hand-build `/checkout` or reservation URLs —
@@ -35,17 +41,18 @@ adjust import paths:
   from the prompt (replace the `<YOUR-CLIENT-ID>` placeholder). The visitor refresh token IS the
   cart identity; it is persisted to localStorage. Do not re-mint anonymously per load or the cart
   silently empties.
-- `src/rest/wix-restaurants.js` — exports, grouped by flow:
-  - **Menu (read-only):** `getFullMenu` (the assembled tree — start here), `listMenus`,
-    `listSections`, `listItems`, `listVariants`, `listModifierGroups`, `listModifiers`, `listLabels`
-  - **Online ordering:** `listOperations`, `getDefaultOperation`, `addItemToCart`,
-    `getCurrentCart`, `updateCartItemQuantity`, `removeFromCart`, `checkout`
-  - **Reservations:** `listReservationLocations`, `getTimeSlots`, `createHeldReservation`,
-    `reserveReservation`
+- `src/rest/wix-restaurants-menu.js` — **Menu (read-only):**
+  `getFullMenu` (the assembled tree — start here), `listMenus`, `listSections`, `listItems`,
+  `listVariants`, `listModifierGroups`, `listModifiers`, `listLabels`
+- `src/rest/wix-restaurants-ordering.js` — **Online ordering:**
+  `listOperations`, `getDefaultOperation`, `addItemToCart`, `getCurrentCart`,
+  `updateCartItemQuantity`, `removeFromCart`, `checkout`
+- `src/rest/wix-restaurants-reservations.js` — **Reservations:**
+  `listReservationLocations`, `getTimeSlots`, `createHeldReservation`, `reserveReservation`
 
 The Menu, Item, ModifierGroup, Operation, Cart, ReservationLocation, TimeSlot, and Reservation
-shapes are documented as JSDoc at the top of `wix-restaurants.js`. Read them before building the
-UI — they describe the key fields and link to the full reference for anything not shown.
+shapes are documented as JSDoc at the top of each helper file. Read the relevant file(s) before
+building the UI — they describe the key fields and link to the full reference for anything not shown.
 
 ## How to wire it (UI is the project's choice)
 - **Menu page** — call `getFullMenu()` once. It returns `{ menus: [{ ...menu, sections: [{ ...section,
